@alamb 2024-04-10
@jackwener  and @jayzhan211  here the next PR in the "don't copy so much in the optimizer"